+++ packages/svelte/src/internal/client/reactivity/deriveds.js
@@ -19,9 +19,7 @@ import {
 	update_reaction,
 	increment_write_version,
 	set_active_effect,
-	component_context,
-	handle_error,
-	get
+	handle_error
 } from '../runtime.js';
 import { equals, safe_equals } from './equality.js';
 import * as e from '../errors.js';
@@ -30,6 +28,7 @@ import { inspect_effects, internal_set, set_inspect_effects, source } from './so
 import { get_stack } from '../dev/tracing.js';
 import { tracing_mode_flag } from '../../flags/index.js';
 import { capture, suspend } from '../dom/blocks/boundary.js';
+import { component_context } from '../context.js';
 
 /**
  * @template V
@@ -39,8 +38,12 @@ import { capture, suspend } from '../dom/blocks/boundary.js';
 /*#__NO_SIDE_EFFECTS__*/
 export function derived(fn) {
 	var flags = DERIVED | DIRTY;
+	var parent_derived =
+		active_reaction !== null && (active_reaction.f & DERIVED) !== 0
+			? /** @type {Derived} */ (active_reaction)
+			: null;
 
-	if (active_effect === null) {
+	if (active_effect === null || (parent_derived !== null && (parent_derived.f & UNOWNED) !== 0)) {
 		flags |= UNOWNED;
 	} else {
 		// Since deriveds are evaluated lazily, any effects created inside them are
@@ -48,16 +51,11 @@ export function derived(fn) {
 		active_effect.f |= EFFECT_HAS_DERIVED;
 	}
 
-	var parent_derived =
-		active_reaction !== null && (active_reaction.f & DERIVED) !== 0
-			? /** @type {Derived} */ (active_reaction)
-			: null;
-
 	/** @type {Derived<V>} */
 	const signal = {
-		children: null,
 		ctx: component_context,
 		deps: null,
+		effects: null,
 		equals,
 		f: flags,
 		fn,
@@ -137,19 +135,14 @@ export function derived_safe_equal(fn) {
  * @param {Derived} derived
  * @returns {void}
  */
-function destroy_derived_children(derived) {
-	var children = derived.children;
+export function destroy_derived_effects(derived) {
+	var effects = derived.effects;
 
-	if (children !== null) {
-		derived.children = null;
+	if (effects !== null) {
+		derived.effects = null;
 
-		for (var i = 0; i < children.length; i += 1) {
-			var child = children[i];
-			if ((child.f & DERIVED) !== 0) {
-				destroy_derived(/** @type {Derived} */ (child));
-			} else {
-				destroy_effect(/** @type {Effect} */ (child));
-			}
+		for (var i = 0; i < effects.length; i += 1) {
+			destroy_effect(/** @type {Effect} */ (effects[i]));
 		}
 	}
 }
@@ -197,7 +190,7 @@ export function execute_derived(derived) {
 
 			stack.push(derived);
 
-			destroy_derived_children(derived);
+			destroy_derived_effects(derived);
 			value = update_reaction(derived);
 		} finally {
 			set_active_effect(prev_active_effect);
@@ -206,7 +199,7 @@ export function execute_derived(derived) {
 		}
 	} else {
 		try {
-			destroy_derived_children(derived);
+			destroy_derived_effects(derived);
 			value = update_reaction(derived);
 		} finally {
 			set_active_effect(prev_active_effect);
@@ -238,9 +231,9 @@ export function update_derived(derived) {
  * @returns {void}
  */
 export function destroy_derived(derived) {
-	destroy_derived_children(derived);
+	destroy_derived_effects(derived);
 	remove_reactions(derived, 0);
 	set_signal_status(derived, DESTROYED);
 
-	derived.v = derived.children = derived.deps = derived.ctx = derived.reactions = null;
+	derived.v = derived.deps = derived.ctx = derived.reactions = null;
 }